Update how queue & playback handle shuffled tracks (#395)
Have the playbackQueue in PlaybackService store the queue in playback order, so that the user can see the order of the queue and upcoming songs in the playback queue component. Add method to Queue to get the playback order.

Also make enqueueAndPlayTracks choose the first track in the queue. This means that when enqueueing tracks from Album/Artist/Genre with shuffle on, it will start playback with a random track. Make a new method enqueueAndPlayTracksFromDoubleClick, used in the tracks collection and track browser components, which keeps the original functionality to start playback with the song that was double clicked, but shuffle the rest.